# services/pi2_assistant/tmp_openapi.json
{"openapi":"3.1.0","info":{"title":"Xubudget API","version":"1.0.0"},"paths":{"/api/healthz":{"get":{"summary":"Health Check","operationId":"health_check_api_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/state":{"get":{"summary":"Get State","operationId":"get_state_api_state_get","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"type":"string","default":"default","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/dashboard_summary":{"get":{"summary":"Get Dashboard Summary","operationId":"get_dashboard_summary_api_dashboard_summary_get","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"type":"string","default":"default","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat":{"post":{"summary":"Chat Endpoint","operationId":"chat_endpoint_api_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessage"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/add_expense":{"post":{"summary":"Add Expense","operationId":"add_expense_api_add_expense_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddExpenseRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/expenses":{"get":{"summary":"Get Expenses","operationId":"get_expenses_api_expenses_get","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"type":"string","default":"default","title":"User Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/categories":{"get":{"summary":"Get Categories","operationId":"get_categories_api_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AddExpenseRequest":{"properties":{"amount":{"type":"number","title":"Amount"},"description":{"type":"string","title":"Description"},"category":{"type":"string","title":"Category"}},"type":"object","required":["amount","description","category"],"title":"AddExpenseRequest"},"ChatMessage":{"properties":{"user_id":{"type":"string","title":"User Id"},"message":{"type":"string","title":"Message"}},"type":"object","required":["user_id","message"],"title":"ChatMessage"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}